MTROPOLIS: Move Print modifier from MTI to Standard
Although MTI had Print in its MTIKIT plugin, most of the content
from MTIKIT, including Print, is found in the EXTRAS and EXPRMNTL
plugins of many mtTropolis 1.1 titles.
