+++ engines/mtropolis/plugin/mti_data.h
@@ -41,18 +41,6 @@ protected:
 	DataReadErrorCode load(PlugIn &plugIn, const PlugInModifier &prefix, DataReader &reader) override;
 };
 
-struct PrintModifier : public PlugInModifierData {
-	PlugInTypeTaggedValue executeWhen;
-	PlugInTypeTaggedValue unknown1Bool;
-	PlugInTypeTaggedValue unknown2Bool;
-	PlugInTypeTaggedValue unknown3Bool;
-	PlugInTypeTaggedValue filePath;
-	PlugInTypeTaggedValue unknown4Bool;
-
-protected:
-	DataReadErrorCode load(PlugIn &plugIn, const PlugInModifier &prefix, DataReader &reader) override;
-};
-
 struct SampleModifier : public PlugInModifierData {
 	PlugInTypeTaggedValue executeWhen;
 	PlugInTypeTaggedValue videoNumber;
